feat(ajout-image): gestion des images: publication, suppression et recupération ...

+++ src/main/java/fr/univrouen/instalite/services/PictureService.java
@@ -1,46 +1,51 @@
 package fr.univrouen.instalite.services;
 
-import java.util.List;
+import java.io.IOException;
+import java.util.Optional;
 
+import fr.univrouen.instalite.entities.Post;
+import fr.univrouen.instalite.utils.MediaStorageUtil;
+import org.apache.commons.io.FilenameUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import fr.univrouen.instalite.dto.PictureRequest;
 import fr.univrouen.instalite.entities.Picture;
-import fr.univrouen.instalite.exceptions.ResourceNotFoundException;
 import fr.univrouen.instalite.repositories.PictureRepository;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public class PictureService {
-    private final PictureRepository pictureRepository;
+    @Autowired
+    private PictureRepository pictureRepository;
 
-    public PictureService(PictureRepository pictureRepository) {
-        this.pictureRepository = pictureRepository;
+    @Autowired
+    private MediaStorageUtil storageUtil;
+
+    public byte[] getFile(String path) throws IOException {
+        return storageUtil.getFile(path);
     }
 
-    public Picture addPicture(PictureRequest pictureRequest) {
-        Picture picture = new Picture();
-        picture.setTitle(pictureRequest.getTitle());
-        picture.setUrl(pictureRequest.getUrl());
-        picture.setPrivate(pictureRequest.isPrivate());
-        return pictureRepository.save(picture);
+    public Optional<Picture> findById(Long id) {
+        return pictureRepository.findById(id);
     }
 
-    public void deletePicture(Long id) {
-        pictureRepository.deleteById(id);
+    public boolean deleteFile(String path,Long id){
+        boolean deleted = storageUtil.deleteFile(path);
+        if(deleted && pictureRepository.existsById(id)){
+            pictureRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public List<Picture> getPublicPictures() {
-        return pictureRepository.findByIsPrivate(false);
-    }
-
-    public List<Picture> getAllPictures() {
-        return pictureRepository.findAll();
-    }
-
-    public Picture updateVisibility(Long id, boolean isPrivate) {
-        Picture picture = pictureRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Picture not found"));
-        picture.setPrivate(isPrivate);
-        return pictureRepository.save(picture);
+    @Transactional
+    public Picture createFile(MultipartFile file, Post post) throws IOException {
+        Picture media = new Picture();
+        String path = storageUtil.createPicture(file);
+        media.setUrl(path);
+        media.setExtension(FilenameUtils.getExtension(media.getUrl()));
+        media.setPost(post);
+        return pictureRepository.save(media);
     }
 }
